--- gribmagic/unity/modules/file_list_handling/remote_file_list_index.py
@@ -1,1 +1,0 @@
-""" functions to scan remote server for file availability """
