+++ pyproject.toml
@@ -1,0 +1,7 @@
+[tool.isort]
+profile = "black"
+src_paths = ["gribmagic", "tests"]
+
+[tool.black]
+line-length = 99
+include = '\.pyi?$'
