+++ tests/unity/download/test_local_store.py
@@ -2,10 +2,10 @@ import os
 import tempfile
 from io import BytesIO
 
-from gribmagic.unity.modules.download.local_store import (
-    bunzip_store,
-    store,
-    tarfile_store,
+from gribmagic.unity.download.decoder import (
+    decode_bunzip,
+    decode_identity,
+    decode_tarfile,
 )
 from tests.unity.fixtures import (
     harmonie_input_file,
@@ -20,12 +20,12 @@ output_file = testdata_path / "output/test.grib2"
 def test_bunzip_store():
     with open(icon_eu_input_file, "rb") as file:
         test_data = file.read()
-    bunzip_store(BytesIO(test_data), output_file)
+    decode_bunzip(BytesIO(test_data), output_file)
     assert output_file.is_file() == True
     os.remove(output_file)
 
 
-def test_store():
+def test_identity_store():
 
     tmpfile = tempfile.NamedTemporaryFile()
     tmpfile.write(b"-" * 42)
@@ -33,7 +33,7 @@ def test_store():
 
     with open(tmpfile.name, "rb") as file:
         test_data = file.read()
-    store(BytesIO(test_data), output_file)
+    decode_identity(BytesIO(test_data), output_file)
     assert output_file.is_file() == True
     os.remove(output_file)
 
@@ -42,6 +42,6 @@ def test_tarfile_store():
     with open(harmonie_input_file, "rb") as file:
         test_data = file.read()
 
-    tarfile_store(BytesIO(test_data), [harmonie_output_file])
+    decode_tarfile(BytesIO(test_data), [harmonie_output_file])
     assert harmonie_output_file.is_file() == True
     os.remove(harmonie_output_file)
